+++ tests/test_project/app_add_not_null_column_followed_by_db_default/models.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+
+from django.db import models
+from django.db.models.functions import Pi
+
+
+class A(models.Model):
+    field = models.IntegerField()
+    not_null_field_db_default = models.IntegerField(
+        null=False,
+        db_default=Pi(),
+    )